Remove unused 'count' param from urls
This makes it easier to glance up and see how much time you've
wasted browsing /r/all, as you can now see the page number in the
URL bar.

I had a grep for count and it doesn't look like anything actually
honours it.  On the desktop site, it is used to add numbers to the
left-hand-side, and is incremented whenever you page forwards or
back. On the mobile version, you have no number beside the posts,
and the 'count' param stays pointlessly set to whatever you type in
(including non-numeric values).